ingestion/requirements.txt
fastapi
uvicorn
minio
kafka-python
pydantic
python-multipart
prometheus_fastapi_instrumentator
requests